Downgraded gulp to v3.9.1.
- Downgraded versions of other dependent packages.
- Discarded "module" type from package.json.
- Renamed *cjs files to *.js.
- Discarded leading "/" to keep consistency with the gulp-generated URL in the templates bundle.
- Performed other necessary changes.
