+++ app/phone-list/phone-list.component.js
@@ -5,7 +5,7 @@ angular.
   module('phoneList').
   component('phoneList', {
     // not using $templateRequest because of the issue described here: https://github.com/risal-shefin/angularJS-phonecat-to-angular/pull/8
-    template: `<ng-include src="'/phone-list/phone-list.template.html'"></ng-include>`,
+    template: `<ng-include src="'phone-list/phone-list.template.html'"></ng-include>`,
     controller: ['phone',
       function PhoneListController(phone) {
         this.phones = [];
