+++ package.json
@@ -2,7 +2,6 @@
   "name": "angular-phonecat",
   "private": true,
   "version": "1.0.0",
-  "type": "module",
   "description": "An experimental application on co-existing AngularJS & Angular",
   "repository": "https://github.com/risal-shefin/angularJS-phonecat-to-angular",
   "license": "MIT",
@@ -45,13 +44,13 @@
     "concat-webpack-plugin": "^0.1.1",
     "cpx": "^1.5.0",
     "css-loader": "^6.9.0",
-    "del": "^6.0.0",
+    "del": "^6.1.1",
     "file-loader": "^6.2.0",
     "glob": "^10.3.10",
-    "gulp": "^4.0.2",
+    "gulp": "^3.9.1",
     "gulp-angular-templatecache": "^3.0.1",
-    "gulp-autoprefixer": "^8.0.0",
-    "gulp-changed": "^4.0.3",
+    "gulp-autoprefixer": "^6.0.0",
+    "gulp-changed": "^3.0.0",
     "gulp-concat": "^2.6.1",
     "gulp-htmlmin": "^5.0.1",
     "gulp-image": "^5.1.0",
@@ -60,7 +59,7 @@
     "gulp-minify-html": "^1.0.6",
     "gulp-rename": "^2.0.0",
     "gulp-sourcemaps": "^3.0.0",
-    "gulp-strip-debug": "^4.0.0",
+    "gulp-strip-debug": "^3.0.0",
     "gulp-uglify": "^3.0.2",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.6.0",
@@ -74,25 +73,29 @@
     "karma-jasmine-html-reporter": "^2.1.0",
     "karma-webpack": "^5.0.0",
     "protractor": "^7.0.0",
+    "run-sequence": "^2.2.1",
     "style-loader": "^3.3.4",
     "terser-webpack-plugin": "^5.3.10",
     "typescript": "~5.2.0",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4"
   },
+  "overrides": {
+    "graceful-fs": "^4.2.11"
+  },
   "scripts": {
     "update-deps": "npm update",
     "prestart": "npm install",
-    "start": "gulp --f gulpfile.cjs && ng serve",
+    "start": "gulp && ng serve",
     "prebuild": "npm install",
-    "build": "gulp --f gulpfile.cjs && ng build",
+    "build": "gulp && ng build",
     "pretest": "npm install",
     "preprotractor": "npm run update-webdriver",
     "preupdate-webdriver": "npm install",
-    "protractor": "protractor e2e-tests/protractor.conf.cjs",
+    "protractor": "protractor e2e-tests/protractor.conf.js",
     "start-http-server": "http-server ./dist/angularjs-phonecat-to-angular -a localhost -p 8000 -c-1",
-    "test": "karma start karma.angularJS.conf.cjs",
-    "test-single-run": "karma start karma.angularJS.conf.cjs --single-run",
+    "test": "karma start karma.angularJS.conf.js",
+    "test-single-run": "karma start karma.angularJS.conf.js --single-run",
     "tsc": "tsc",
     "tsc:w": "tsc -w",
     "update-webdriver": "webdriver-manager update"
